requirements.txt
graphviz==0.20.1
ipython==7.31.1
matplotlib==3.5.2
numpy==1.23.5
openpyxl==3.0.10
pandas==1.4.4
Pillow==9.5.0
plotly==5.9.0
pydotplus==2.0.2
scikit_learn==1.2.2
seaborn==0.11.2
streamlit==1.22.0

